requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
mysql-connector-python==9.4.0
mysqlclient==2.2.7
python-dotenv==1.1.1
requests==2.32.4
urllib3==2.5.0
